+++ x/slashing/keeper/rank_test.go
@@ -104,7 +104,6 @@ func createValidators(t *testing.T, app *simapp.SekaiApp, ctx sdk.Context, accNu
 		require.NoError(t, err)
 
 		validator, err := stakingtypes.NewValidator(
-			sdk.NewDec(1234),
 			valAddr,
 			pubKey,
 		)