+++ x/staking/abci_test.go
@@ -25,7 +25,6 @@ func TestItUpdatesTheValidatorSetBasedOnPendingValidators(t *testing.T) {
 	require.NoError(t, err)
 
 	validator1, err := stakingtypes.NewValidator(
-		types.NewDec(1234),
 		valAddr1,
 		pubKey,
 	)
@@ -108,7 +107,6 @@ func TestItRemovesFromTheValidatorSetWhenInRemovingQueue(t *testing.T) {
 			require.NoError(t, err)
 
 			validator1, err := stakingtypes.NewValidator(
-				types.NewDec(1234),
 				valAddr1,
 				pubKey,
 			)
@@ -192,7 +190,6 @@ func TestItIncludesItBackToValidatorSetOnceReactivatingIt(t *testing.T) {
 			require.NoError(t, err)
 
 			validator1, err := stakingtypes.NewValidator(
-				types.NewDec(1234),
 				valAddr1,
 				pubKey,
 			)