+++ x/staking/client/cli/flags.go
@@ -18,7 +18,6 @@ const (
 // AddValidatorFlags adds the flags needed to create a validator.
 func AddValidatorFlags(cmd *cobra.Command) {
 	cmd.Flags().String(FlagMoniker, "", "the Moniker")
-	cmd.Flags().String(FlagComission, "", "the commission")
 	cmd.Flags().String(FlagValKey, "", "the validator key")
 	cmd.Flags().String(cli.FlagPubKey, "", "the public key")
 }