+++ x/staking/keeper/msg_server.go
@@ -49,16 +49,12 @@ func (k msgServer) ClaimValidator(goCtx context.Context, msg *types.MsgClaimVali
 	}
 
 	moniker := strings.Trim(msg.Moniker, " ")
-	if len(moniker) > 64 {
-		return nil, types.ErrInvalidMonikerLength
-	}
-
 	_, err = k.keeper.GetValidatorByMoniker(ctx, moniker)
 	if err == nil {
 		return nil, types.ErrValidatorMonikerExists
 	}
 
-	validator, err := types.NewValidator(msg.Commission, msg.ValKey, pk)
+	validator, err := types.NewValidator(msg.ValKey, pk)
 	if err != nil {
 		return nil, err
 	}
@@ -66,18 +62,20 @@ func (k msgServer) ClaimValidator(goCtx context.Context, msg *types.MsgClaimVali
 	k.keeper.AddPendingValidator(ctx, validator)
 
 	// register identity record moniker field when claiming a validator
-	k.govKeeper.RegisterIdentityRecords(ctx, sdk.AccAddress(validator.ValKey), []govtypes.IdentityInfoEntry{
+	err = k.govKeeper.RegisterIdentityRecords(ctx, sdk.AccAddress(validator.ValKey), []govtypes.IdentityInfoEntry{
 		{
 			Key:  "moniker",
 			Info: moniker,
 		},
 	})
+	if err != nil {
+		return nil, err
+	}
 
 	ctx.EventManager().EmitEvent(
 		sdk.NewEvent(
 			types.EventTypeClaimValidator,
 			sdk.NewAttribute(types.AttributeKeyMoniker, msg.Moniker),
-			sdk.NewAttribute(types.AttributeKeyCommission, msg.Commission.String()),
 			sdk.NewAttribute(types.AttributeKeyValKey, msg.ValKey.String()),
 			sdk.NewAttribute(types.AttributeKeyPubKey, msg.PubKey.String()),
 		),